version_0.2-yoomoney Build #5: Рефакторинг кода
