+++ handlers/channels/__init__.py
@@ -1,1 +1,1 @@
-from . import to_user
+from . import to_users
