+++ handlers/errors/error_handler.py
@@ -11,7 +11,6 @@ from loader import dp
 async def errors_handler(update, exception):
     """
     Exceptions handler. Catches all exceptions within task factory tasks.
-    :param dispatcher:
     :param update:
     :param exception:
     :return: stdout logging
