+++ handlers/users/__init__.py
@@ -1,11 +1,8 @@
 from . import cancel
 from . import help
 from . import start
-from . import instagram_post
-from . import instagram_highlight
-from . import instagram_story
-from . import instagram_user
+from . import subscriber
+from . import not_subscriber
 from . import info
 from . import subscribe
 from . import referral
-# from . import terms
