+++ handlers/users/subscriber/instagram_post.py
@@ -1,5 +1,4 @@
 from aiogram.types import Message, InlineQuery
-from keyboards.inline.generate import subscribe_keyboard
 from loader import dp, bot, upload_client
 from utils.instagram import InstagramPost
 
@@ -9,13 +8,6 @@ async def instagram_post_handler(message: Message):
     post = InstagramPost(message.text)
     await post.start()
     await post.send_to(bot=bot, upload_client=upload_client, chat_id=message.chat.id)
-
-
-@dp.message_handler(instagram_post=True, state='*')
-async def not_instagram_post_handler(message: Message):
-    await message.answer(
-        text='🤖 Подключи подписку и скачивай контент из <pre>Instagram</pre> анонимно и без ограничений\nХочешь пользоваться ботом бесплатно? /referral\nХочешь подключить подписку? Кликай на кнопку ниже!',
-        reply_markup=subscribe_keyboard())
 
 
 @dp.inline_handler(is_subscriber=True, instagram_inline_post=True, state='*')
